Update submodule biosoup

+++ src/main.cpp
@@ -13,7 +13,7 @@
 
 #include "ram/minimizer_engine.hpp"
 
-std::atomic<std::uint64_t> biosoup::Sequence::num_objects{0};
+std::atomic<std::uint32_t> biosoup::Sequence::num_objects{0};
 
 namespace {
